Collapsed NVIDIA Riva modules into one file for better compat with unit tests.
