--- templates/nvidia_riva/nvidia_riva/__init__.py
@@ -1,3 +1,0 @@
-from nvidia_riva.chain import chain
-
-__all__ = ["chain"]
